Move Figi string allocation into `from_str`
Here, parse_figi returns a string slice (&'s str) that borrows from the input string 's. The lifetime 's indicates that the returned string slice has the same lifetime as the input string. In other words, the returned string slice is a borrowed reference to a portion of the input string.
When we call parse_figi.parse(s) in the from_str method, it returns a Result<&str, _>. If the parsing is successful, we get a borrowed string slice &str. We then use s.to_owned() to create a new owned String from that borrowed string slice. This is where the allocation happens, as we are creating a new owned string to store in the Figi newtype.

The key difference is where the allocation happens. In the first approach, the allocation happens only when creating the Figi newtype, which is necessary because we need to store an owned string inside the struct. In the second approach, the allocation happens inside parse_figi, which is unnecessary because we can simply return a borrowed string slice and postpone the allocation until we actually need to create the Figi newtype.